Adding routing to search app
