+++ search/src/Components/register.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function RegisterComponent() {
+    return (
+        <h1>Register Page</h1>
+    )
+}
+
+export default RegisterComponent;
